requirements.txt
pandas
folium==0.19.2
streamlit-aggrid==1.0.5
streamlit_folium==0.23.2
plotly_express==0.4.0
networkx==3.4.2
pyvis==0.3.2